requirements.txt
faiss_cpu==1.7.4
Flask==2.3.2
Flask_Cors==3.0.10
modAL==0.4.1
modAL==0.4.1
nltk==3.8.1
numpy==1.24.3
pandas==2.0.1
scikit_learn==1.2.2
txtai==5.5.1
gunicorn
